# Conflicts:
#	docs/DeveloperGuide.md
#	docs/UserGuide.md
#	src/main/java/ui/Output.java

+++ src/main/java/workouts/GymStation.java
@@ -8,6 +8,7 @@ import java.util.ArrayList;
  * Represents a GymStation object.
  */
 public class GymStation {
+    //@@author JustinSoh
     protected String stationName;
     protected ArrayList<GymSet> sets = new ArrayList<>();
     protected int numberOfSets;
@@ -21,7 +22,7 @@ public class GymStation {
      * @param numberOfRepetitions Number of reps done.
      * @param weightsList         Weight used.
      */
-    protected GymStation(String name, int numberOfSets, int numberOfRepetitions, ArrayList<Integer> weightsList) {
+    protected GymStation(String name, int numberOfSets, int numberOfRepetitions, ArrayList<Double> weightsList) {
         this.stationName = name;
         this.numberOfSets = numberOfSets;
         processSets(weightsList, numberOfRepetitions);
@@ -33,7 +34,7 @@ public class GymStation {
      * @param weightsList     The weight done for the particular set.
      * @param numberOfRepetitions The number of repetitions done for the particular set.
      */
-    public void processSets(ArrayList<Integer> weightsList, int numberOfRepetitions) {
+    public void processSets(ArrayList<Double> weightsList, int numberOfRepetitions) {
         for (int i = 0; i < numberOfSets; i++) {
             GymSet newSet = new GymSet(weightsList.get(i), numberOfRepetitions);
             sets.add(newSet);
